+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/onboarding/persistence/AnswerRepository.java
@@ -1,15 +1,18 @@
-package org.sopt36.ninedotserver.onboarding.repository;
+package org.sopt36.ninedotserver.onboarding.persistence;
 
 import java.util.List;
-import org.sopt36.ninedotserver.onboarding.domain.Answer;
-import org.sopt36.ninedotserver.onboarding.domain.Domain;
+import org.sopt36.ninedotserver.onboarding.model.Answer;
+import org.sopt36.ninedotserver.onboarding.model.Domain;
+import org.sopt36.ninedotserver.onboarding.persistence.querydsl.AnswerRepositoryCustom;
+import org.sopt36.ninedotserver.onboarding.port.out.AnswerRepositoryPort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnswerRepository extends JpaRepository<Answer, Long>, AnswerRepositoryCustom {
+public interface AnswerRepository extends JpaRepository<Answer, Long>, AnswerRepositoryPort,
+    AnswerRepositoryCustom {
 
     @Query("SELECT a.content FROM Answer a WHERE a.user.id = :userId AND a.question.domain = :domain")
     List<String> findAllAnswerContentsByUserId(
