+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/onboarding/persistence/querydsl/AnswerRepositoryImpl.java
@@ -1,7 +1,4 @@
-package org.sopt36.ninedotserver.onboarding.repository;
-
-import static org.sopt36.ninedotserver.onboarding.domain.QAnswer.answer;
-import static org.sopt36.ninedotserver.onboarding.domain.QQuestion.question;
+package org.sopt36.ninedotserver.onboarding.persistence.querydsl;
 
 import com.querydsl.core.Tuple;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -9,9 +6,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.sopt36.ninedotserver.onboarding.domain.Domain;
-import org.sopt36.ninedotserver.onboarding.domain.QAnswer;
-import org.sopt36.ninedotserver.onboarding.domain.Question;
+import org.sopt36.ninedotserver.onboarding.model.Domain;
+import org.sopt36.ninedotserver.onboarding.model.QAnswer;
+import org.sopt36.ninedotserver.onboarding.model.QQuestion;
+import org.sopt36.ninedotserver.onboarding.model.Question;
 
 @RequiredArgsConstructor
 public class AnswerRepositoryImpl implements AnswerRepositoryCustom {
@@ -19,7 +17,7 @@ public class AnswerRepositoryImpl implements AnswerRepositoryCustom {
     private final JPAQueryFactory queryFactory;
 
     public List<Question> findQuestionsByUserIdAndDomain(Long userId, Domain domain) {
-        QAnswer a = answer;
+        QAnswer a = QAnswer.answer;
 
         return queryFactory
             .select(a.question)
@@ -34,16 +32,16 @@ public class AnswerRepositoryImpl implements AnswerRepositoryCustom {
     @Override
     public Map<String, String> findQnAMapByUserId(Long userId) {
         List<Tuple> results = queryFactory
-            .select(question.content, answer.content)
-            .from(answer)
-            .join(answer.question, question)
-            .where(answer.user.id.eq(userId))
+            .select(QQuestion.question.content, QAnswer.answer.content)
+            .from(QAnswer.answer)
+            .join(QAnswer.answer.question, QQuestion.question)
+            .where(QAnswer.answer.user.id.eq(userId))
             .fetch();
 
         return results.stream()
             .collect(Collectors.toMap(
-                tuple -> tuple.get(question.content),
-                tuple -> tuple.get(answer.content)
+                tuple -> tuple.get(QQuestion.question.content),
+                tuple -> tuple.get(QAnswer.answer.content)
             ));
     }
 }
