+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/onboarding/persistence/querydsl/ChoiceRepositoryCustom.java
@@ -1,0 +1,9 @@
+package org.sopt36.ninedotserver.onboarding.persistence.querydsl;
+
+import java.util.List;
+import org.sopt36.ninedotserver.onboarding.model.Choice;
+
+public interface ChoiceRepositoryCustom {
+
+    List<Choice> findJobList();
+}
