--- src/main/java/org/sopt36/ninedotserver/onboarding/repository/ChoiceRepositoryCustom.java
@@ -1,9 +1,0 @@
-package org.sopt36.ninedotserver.onboarding.repository;
-
-import java.util.List;
-import org.sopt36.ninedotserver.onboarding.domain.Choice;
-
-public interface ChoiceRepositoryCustom {
-
-    List<Choice> findJobList();
-}
