--- src/main/java/org/sopt36/ninedotserver/onboarding/repository/QuestionRepository.java
@@ -1,14 +1,0 @@
-package org.sopt36.ninedotserver.onboarding.repository;
-
-import java.util.List;
-import org.sopt36.ninedotserver.onboarding.domain.Domain;
-import org.sopt36.ninedotserver.onboarding.domain.Question;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-
-    List<Question> findAllByActivatedTrueAndDomain(Domain domain);
-
-}
